fix: allow CredentialCache.Retrieve to return before initialization (#50748)
* fix: allow CredentialCache.Retrieve to return before initialization

Fixes #50747

This PR adds an option to allow CredentialsCache.Retrieve to return with
an error before CredentialsCache.SetGenerateOIDCTokenFn has been called.
We have been relying on this behaviour for Auth service startup, because
the credential cache is a dependency for the audit storage, which is a
dependency for the auth server, which is a dependency for the credential
cache's GenerateOIDCTokenFn.
Some non-critical checks during auth startup must fail instead of
hanging on blocking the process from successfully starting.

* use default io timeout

* synchronize access to generateOIDCTokenFn