[chore] replace gitlab client with `gitlab.com/gitlab-org/api/client-go` (#5811)
GitLab has adopted the Go client previously hosted at `github.com/xanzy/go-gitlab` and deprecated it in favor of `gitlab.com/gitlab-org/api/client-go`. The codebase remains unchanged as it is a direct fork of the original client, but it is now maintained under GitLab's organization.

Signed-off-by: Tiago Silva <tiago.silva@goteleport.com>

Source-commit: gravitational/teleport.e@1bc4a6909732d3b8b98b19fab56d9a39f228ec01
Source-pull-requests: gravitational/teleport.e#5811