[v16] buildbox: Update centos7-assets buildbox for EOL (#50750)
Update the centos7-assets buildbox to build with `vault.centos.org`
instead of `mirror.centos.org` as the latter has gone away now centos7
is EOL. This same change was made to the centos7 buildbox back in July
but was not done to the assets buildbox which rarely needs to be
rebuilt. However a v15 clang change that needs to be applied has shown
this change was missed.

This was done to the centos7 buildbox in #43785 (0f4eadd8d0a3).