Consistently use logutils.TypeAttr (#5655)
When converting to slog there have been various mechanisms used to
include an objects type in a slog attribute to replace %T from a
logrus message. This unifies all type attribute to use
logutils.typeAttr instead of calling reflect/fmt directly. The
former ensures that the type is only calculated if the message is
going to be logged.

Source-commit: gravitational/teleport.e@3e8af2e980f1fc05c31e4f015f5709316887d7f0
Source-pull-requests: gravitational/teleport.e#5655