Complete NewIdentityService renaming process (#5832)
The function signature of the identity service was changed in
https://github.com/gravitational/teleport/pull/40941 and a new V2
variant was added to allow the change without breaking teleport.e.

Now that everything has been converted to use V2, the original
function signature was changed, however the follow up to remove
the V2 variant was neglected. This converts all calls in teleport.e
to prefer the original variant so that V2 may finally be deleted.

Source-commit: gravitational/teleport.e@6831bf6d753367225b6bda5a01189e3065367c9d
Source-pull-requests: gravitational/teleport.e#5832