Fix incorrectly named dynamo events scaling policy (#50990)
The read scaling policy name was incorrectly changed to match the
write scaling policy. This prevents upgrading from a v16 cluster
with a dynamo audit backend configured to use autoscaling to v17.
To resolve, when conflicts are found the incorrectly named scaling
policy is removed, and replaced by the correctly named one.