Fix some Vars
Fix another instance

In general, add a Var if it's a step-by-step guide that would benefit
from copying pasting. For examples in conceptual guides, use a
placeholder.

In some cases, the only instance of a Var is in a configuration block,
but is the only Var in that block. In that case, replace it with a
placeholder value to be consistent with other config fields.