Hide features if dashboard tenant (#5854)
This PR makes an update to our logic to the "backdoor" that would _keep_
things hidden after making updates to show most of the features to
promote discoverability. Currently, we only continuted to hide them if
their license specified it, but this incorrectly ignored dashboard,
usage based, and team tenants as well.

Source-commit: gravitational/teleport.e@14bdca457653e2309cee35ab1da9b7628795392d
Source-pull-requests: gravitational/teleport.e#5854