Hide specific features from nav is dashboard tenant (#51118)
This PR makes an update to our logic to the "backdoor" that would keep
things hidden after making updates to show most of the features to
promote discoverability. Currently, we only continued to hide them if
their license specified it, but this incorrectly ignored dashboard,
usage based, and team tenants as well.