Multi-domain logins for Windows Desktop
When connecting to a Windows desktop, Teleport has always
assumed that the Windows user we are connecting as belongs
to the same Active Directory Domain as the Windows host we
are trying to connect to. This prevents some setups from
working with Teleport.

We fix this by looking at the Windows login, and if it looks
like a full UPN (ie like an email address) then we use the
domain from the UPN rather than assuming it matches the
domain the desktop is registered in.

In order for this to work, Teleport needs to be able to make
an LDAP query for the user's SID in different AD domains
(and therefore to different LDAP servers). Teleport achieves
this by querying its default LDAP server and then following
any LDAP referrals that are returned. This means that:
- LDAP referrals must be set up correctly
- The service account that Teleport is using for LDAP must
  have permission to query multiple domains.
- The windows_desktop_service must have network connectivity
  to the LDAP servers in multiple domains.

Co-authored-by: Zac Bergquist <zac.bergquist@goteleport.com>