Improve the tests for listing active sessions (#51246)
Use a more standard table-driven test setup, as the previous
approach put the description of the test at the very end.

Also adds a bit more flexibility to allow for additional more
complicated tests to be added, and adds one additional test
case to verify that explicit deny rules work.