Upgrade vite to 6.0.11
This will upgrade vite. I'm getting an "unmet dep" warning for
electron-vite saying it expects a max of vite 5, however we are
currently on the latest version of electron-vite so I don't think there
is much we can do here. We have already been running vite 6+ for months
now tho and electron-vite has been working fine so I think its ok to
ignore for now.