[vnet] windows ip and route configuration
This PR adds OS configuration for VNet on Windows.
Specifically, the TUN interface is now configured with a V6 and V4 IP
address, and IP routes are configured so that IP packets in the VNet IP
ranges for each cluster are routed to the TUN interface and handled by
VNet.

This PR does *not* configure the VNet DNS nameserver on Windows, that
will come in a following PR.

With these changes, VNet kind of works, without DNS. You can manually
query the IP address of VNet's DNS server and get back a v4 and v6
address for the app. TCP connections to either of these addresses then
work for connecting to the app.