[v17] kube: properly return the reason for connection disruption (#51454)
* kube: properly return the reason for connection disruption (#51398)

* kube: properly return the reason for connection disruption

There are several cases where connection monitor can terminate an
ongoing connection. Iddle timeout, certificate expiring among others are
some reasons for the connection to be terminated.

For Kubernetes access, the underlying error is never propagated back to
the client so they don't receive the reason for the exec session being
terminated.

This PR fixes that by adding an hook to write the client error response
into the connection error channel for clients to be aware.

Part of https://github.com/gravitational/teleport/issues/18496

* handle review comments

* handle review comments

* fix slog ref