fix: fix potencial data race during `InitLoggerForTests` calls
`utils.InitLoggerForTests` calls `flag.Parse` to check if the tests were
configured with the verbose mode enabled.

This approach is required because in most initializations a `TestMain`
is invoked prior to execute the tests and `InitLoggerForTests` is called
under this umbrella.

In go testing framework, `flag.Parse` is only called at the top of
`testing.M.Run` invocation - called by `TestMain` if any or by the
default `TestMain` function.

The data race is introduced if a go test executes
`utils.InitLoggerForTests` without calling the same function or
`NewSlogLoggerForTests` in `TestMain` function. This occurs because go
testing framework already parsed the values and a subsequent parse using
the global flag parser - we run the parse in a sync.Once but it doesn't prevent other code to call
flag.Parse - changes the variables cached in `testing` package, leading
to data races.

The fix is to check if flag.Parse was previously invoked either by our
custom `TestMain` function that calls `InitLoggerForTests` or by
`testing.M.Run` call. In any case, if the test package doesn't include a
`TestMain` function that calls `InitLoggerForTests` but one of its tests
does call `NewSlogLoggerForTests`, no data race will occur because test
cases only run after `testing.M.Run` invocation.

Signed-off-by: Tiago Silva <tiago.silva@goteleport.com>