Emphasize exec in the browser
Updates the Kubernetes cluster connect dialog to promote exec in
the browser over the CLI instructions. When this was originally
rolled out the intention was to place less emphasis on exec in the
browser, however, that sentiment changed with Postgres connect in
the browser. This updates the dialog to match the ordering and
copy of the Kubernetes dialog to match the Postgres equivalent.