Polish Kubernetes and Postgres connection dialogs (#51554)
This aims to add a few additional touches and unify the modals
displayed in the web ui in response to clicking Connect for a
kubernetes cluster or a database.

Both prompts now present connecting in the web ui as the primary
option. The CLI text is secondary, and the Postgres CLI instructions
now have the same `Or connect in the CLI using tsh` copy that the
Kubernetes prompt did.

The wording was unified across the two prompts where appropriate
and the headings were tweaked to include resource information instead
of a generic Connect to Kubernetes/Postgres title.