Polish Kubernetes and Postgres connection dialogs (#51507) (#51557)
This aims to add a few additional touches and unify the modals
displayed in the web ui in response to clicking Connect for a
kubernetes cluster or a database.

The wording was unified across the two prompts where appropriate
and the headings were tweaked to include resource information instead
of a generic Connect to Kubernetes/Postgres title.