Machine ID: Support path-based Kubernetes routing (#50898) (#51535)
* Machine ID: Support path-based Kubernetes routing

This adds a new `kubernetes/v2` service to support path-based
routing, which allows clients to access an arbitrary number of
Kubernetes clusters using a single issued identity.

It can be used with `tbot start kubernetes/v2` and specifying one or
more explicit clusters with `--kubernetes-cluster-name` or a label
selector with `--kubernetes-cluster-labels`.

* Added tests for new Kubernetes selector behavior

* Fix failing tests

* Fix test failure after rebase

* Rename CLI parameters to match workload ID selector syntax

* Fix broken test after flag rename

* Remove unused function

* Remove user-configurable roles from kubernetes/v2