DiscoveryConfigStatus: update counters once per iteration
DiscoveryService can process DiscoveryConfig matchers and will update
its Status field depending on the number of found resources.

Currently, it was updating them in real-time-ish: every time it found a
new resource, it would update the counter in memory and propagate that
change into the cluster shared storage.

However, this resulted in counters flutuating a lot and causing
confusion from the user's point of view (being at 0 every couple of
minutes).

Instead, this PR, ensures we update the values only once per iteration
cycle (happens at discovery_service.poll_interval - defaults to 5
miunutes).
This means we lose real-time observation of the counters, but we gain
stability on the values which should help UX.