[vnet] refactor osconfig (#51577)
* [vnet] refactor osconfig

This PR refactors VNet's OS configuration code.
There is now a `osConfigurator` struct that runs the OS configuration
loop and calls out to a `targetOSConfigProvider` which determines the
current target OS configuration.
This `targetOSConfigProvider` can be implemented for the MacOS daemon
process by moving some of the existing code around.
In a following PR I will implement it for Windows which will call out to
the user process to get the current target OS configuration.

This PR doesn't really change any of the logic that runs on MacOS, it
just refactors things so that we can swap in a different implementation
for Windows.

FYI, after I get the Windows code fully working, I plan to migrate the
MacOS code to follow the new way Windows handles things, where the admin
process handles all networking and calls out to the user process for
everything that uses a Teleport client.
This will unify multiple codepaths and eliminate the
profileOSConfigProvider added here.

* fix lint on linux

* use ip-aware function to compute TUN address