Simplify desktop canvas
The main changes are
- simplify the relationship between DesktopSession and TdpClientCanvas.
This moves some logic around and tries to consolidate canvas only
related functionality to useTdpClientCanvas and the rest into
useDesktopSession

- Remove the need to add listeners for tdpclient events and instead,
apply eventHandlers directly.

Notes from previous PRs before the squash/rebase

Add the eventHandlers type to the client so it can  be used instead of
anemitter and also instantiate the client in `useDesktopSession`
insteadof the `useTdpClientCanvas`

Move client calls to `useDesktopSession`. `useTdpClientCanvas`
shouldn'tset anything when it comes to client callbacks. Really, the
only thing`useTdpClientCanvas` should be is a glorified object of canvas
callbacksthat we can spread as props into TdpClientCanvas.

From the looks of it, we still need it in `useDesktopSession` because
ofthe canvasOnPngFrame and canvasOnnBmpFrame methods that call a draw
tothe canvas. We basically return those methods from
`useTdpClientCanvas`and then call them in the `useDesktopSession`
`onPngFrame` etc.

Add typesafe eventHandler update method. We can set any number of
handlersin the constructor, but this method will allow us to set (or
update) eventhandlers during the lifetime of the client. It's also type
safe so thecallsite knows exactly which handler keys can be added and
their signature