Allow the proxy to prefer an in-process desktop service
Add an opt-in env var that tells the proxy to prefer a desktop service running in-process
as an alternative to selecting a service at random. This can benefit self-hosted deployments
where the proxy and desktop agent run in the same process by preventing the proxy from
selecting an agent that may be in another geo over one that's available on the same host.

This is a crude and specialized alternative to a more general latency-based dialing strategy
(which is still something we'd like to do in the future).

Updates #51879
Updates #40905