Add `OnProvisioned` callback to SCIM provisioner (#6032)
* Add `OnProvisioned` callback to SCIM provisioner

The Identity Center integration currently has a race when a Teleport user is cretad: both the
user provisioning system and account assignmnt provisioning system immediately attempt to to
their jobs, resulting in a situation where the assignment provisioner fails because the user
it wants to operate on is not yet created.

This patch adds an `OnProvisioned` event callback to the SCIM provisioner configuarion that
will be invoked when a user transitions to the PROVISIONED state. This will eventually be
picked up (in a subsequent PR) by the Identity Center integration, which will use it as a
trigger to retry provisioning the user's account assignments.

Split out from #5885, which will contain the IC part of the change

* address review feedback

Source-commit: gravitational/teleport.e@8cf9451b61361ddfd6016e8f97e52f903f0d1545
Source-pull-requests: gravitational/teleport.e#6032