[v15] build: Fix buildbox for fips builds of ironrdp
Apply some changes that have been made to the later branches to v15 to
try to get the FIPS release builds working. It seems to have broken with
an ironrdp update recently.

This rolls back clang in the buildbox from v14 to v12 as required by
FIPS compliance, and also disables rdpclient for arm64 FIPS builds
(although that is actually controlled by the e/Makefile, so this change
here is just to keep things in sync).