Add Scheduled Upgrade window to Help & Support page (#6062)
This is a partial revert of
https://github.com/gravitational/teleport.e/pull/5567

The UX team still desires the upgrade window form to live on the Manage
Clusters page so this just adds to back to where it was without removing
it from the manage clusters page.

A future PR will update the RBAC resource that allows access in a more
clear way to the manage clusters page.

Source-commit: gravitational/teleport.e@85829f614fc789da87b453a093b244fa064f245c
Source-pull-requests: gravitational/teleport.e#6062