Clean up discover alerts (#52204)
Discover was using a custom HintBox component instead of the in-page
alerts that are part of our design system. As a result, some items
rendered without enough contrast. Use the new alert component to
ensure consistency across the UI.

This also fixes a styling but where links inside alerts were not
rendered with the correct text color.

Co-authored-by: Rafał Cieślak <rafal.cieslak@goteleport.com>