Address race condition in IC when creating Ephemeral Users (#5885)
* Address race condition in IC when creating Ephemeral Users

There is a race conditon in the Identity Center integration where creating
a user causes the user's AWS account assignments to sometimes be provisioned
before the user itself if fully provisioned, causing the account assignment
provisioing to fail.

Building on the events introduced in #6032, this patch adds an forces a re-
provision of a User or Access List's Account Assignments when the principal
has been reprovisioned via the SCIM provisioner.

This patch also adds a fallback to External ID lookups, where the lookup will
try the cache first and fall back to the main service if the External ID isn't
found. This is because we sometimes need to look up a principal's External ID
immediately after learning it, often before the provisioning state resource
write containing the new External ID has been propagated to the cache.

Most of these issues are mitigated when users are created off-line, as the
periodic synchronization in the IC integration will clean up the failed
provisioning before the users have a chance to log in and notice the missing
permissions. It is a more obvious problem for the ephemeral users cretaed by
SAML connectors, where a user will be using Teleport immediately after they
are created.

* Apply suggestions from code review

Co-authored-by: Pawel Kopiczko <pawel.kopiczko@goteleport.com>

* Re-arrange IC construction to avoid needing event setters

* Refactor External ID lookup fallback

* Remove redundant copy of `AssertSequenceLength()`

* Update service logging

* Tidy up assertion

---------

Co-authored-by: Pawel Kopiczko <pawel.kopiczko@goteleport.com>

Source-commit: gravitational/teleport.e@061fc6c401e7330a00ee7e1e2bff92635eaf8d45
Source-pull-requests: gravitational/teleport.e#5885