Implement buttons in VNet diag report (#52346)
* Remove unused AppConnectionItem

This was implemented back when we were thinking of showing recent VNet
connections in the panel. We've never had time to implement this, so
it's better to just remove this rather keeping this component in the
project.

* Implement button for re-running diagnostics

* Add reportToText

* Implement button for copying report to clipboard

* Clean up JSDocs for main process client

JSDocs for individual methods should be defined on the type, so that the
JSDocs are available in any place that uses the type, rather than only
where we use the implementation of the type.

* Implement button for saving report to file

* Add JSDoc for reportToText

* Simplify created at in text report