Remove PAM user creation deprecation (#52270)
The deprecation often leads to confusion and support load as
suggesting PAM as a solution to problems is met with questions
about why a customer should use a solution that will be
deprecated in the future.

While Teleport's built in host user provisioning may be the
preferred approach to create local users on the fly, folks that
already leverage the PAM stack may prefer PAM to provision users.