Remove the Auto-Discovery reference index page
Closes #52266

A link to the Auto-Discovery reference from the category index page
incorrectly points to the main Reference section. This is because the
docs engine appears to rewrite a link to `reference/reference.mdx` from
`auto-discovery.mdx` to point to the incorrect page. The error takes
place regardless of whether the link is generated from `remark-toc` or
added manually.

The link generation issue takes place infrequently, so we can address it
with a workaround until the issue occurs more regularly and we can
gather more context. This change removes the Auto-Discovery reference
page, since the page is a list of two links, one of which points to a
different section of the docs.