Coalesces Identity Center events during handling (#6077)
* Address race condition in IC when creating Ephemeral Users

There is a race conditon in the Identity Center integration where creating
a user causes the user's AWS account assignments to sometimes be provisioned
before the user itself if fully provisioned, causing the account assignment
provisioing to fail.

Building on the events introduced in #6032, this patch adds an forces a re-
provision of a User or Access List's Account Assignments when the principal
has been reprovisioned via the SCIM provisioner.

This patch also adds a fallback to External ID lookups, where the lookup will
try the cache first and fall back to the main service if the External ID isn't
found. This is because we sometimes need to look up a principal's External ID
immediately after learning it, often before the provisioning state resource
write containing the new External ID has been propagated to the cache.

Most of these issues are mitigated when users are created off-line, as the
periodic synchronization in the IC integration will clean up the failed
provisioning before the users have a chance to log in and notice the missing
permissions. It is a more obvious problem for the ephemeral users cretaed by
SAML connectors, where a user will be using Teleport immediately after they
are created.

* Apply suggestions from code review

Co-authored-by: Pawel Kopiczko <pawel.kopiczko@goteleport.com>

* Re-arrange IC construction to avoid needing event setters

* Refactor External ID lookup fallback

* Remove redundant copy of `AssertSequenceLength()`

* Update service logging

* Tidy up assertion

* Coalesces Identity Center events during handling

In ordrer to reduce the rework done by the Identity Center integration, this patch
changes how the integration handles events from its resource monitor. Prior to this
patch it would process each event in series. This would result in lots of rework in
situations that trigger full recalculations, or where a single event triggers
multiple events on a single principal (for example, editing Access List memberships).

This patch changes the event handling such that it now
 - Processes the events in batches
 - Coalesces all events for the same principal in the batch down to one sigle event
 - If any of the events in the batch trigger a full recalculation, all other events
   are discarded, because their target principals will all be re-calculated anyway
   during the full recalc

* Add todo comments on  impementation

* Extract the IC event batch reader into its own functon

* Batch by time interval

* Updates from editor

* Apply feedback

* test fixup

* Updates from editor

---------

Co-authored-by: Pawel Kopiczko <pawel.kopiczko@goteleport.com>

Source-commit: gravitational/teleport.e@b45007ab0689ad387dd830bc116ea5b239c4084f
Source-pull-requests: gravitational/teleport.e#6077