Refactor client canvas (#52431)
* Remove non-canvas related listeners from `TdpClientCanvas`

* Simplify event handlers setup by using React handlers

* Listen to canvas resize events instead of window ones

* Expose imperative API to set pointer

* Expose imperative API to render frames

* Expose imperative API to change canvas resolution

Also, we no longer need to set the canvas size manually; it is now handled by `objectFit: scale-down`.

* Remove duplicate player shutdown

* Expose imperative API to clear canvas

* Remove unnecessary display size syncing

We get the canvas resolution from the server, there doesn't seem to be a need to calculate it from the browser window size.

* Fix focusing canvas

It didn't work because it happened too early, before we even showed the MFA dialog which stole focus.
This is a temporary fix, I will do it more elegantly in the future.

* Minor code changes

Remove component memoization, it didn't work anyway and doesn't seem needed. We don't have anything particularly expensive to render here from React's perspective.

* Add a TODO for improving rendering functions