Standardise "upstream" and "downstream" usage in provisioner (#6104)
The provisioning code refers to the remote system as "downstream", from
the notion that provisioning information flows from Teleport downstream to
the remote system.

Some of the test code uses "upstream" to refer to the same thing.

This patch updates the test code useage to match the primary code.

Source-commit: gravitational/teleport.e@4149d9502ce66131a72949458b865ff96b6a8076
Source-pull-requests: gravitational/teleport.e#6104