Remove the check-redirects script
Backports #52750

The `docs/check-redirects` code was intended to run as part of an
automated workflow. But because the workflow would need to check private
repos as well as public ones, it was not possible to run this code from
`gravitational/teleport`. In addition, we now use a workflow that is
stored in the `shared-workflows` repo to check redirects.