Add iteration support to dynamo (#52517)
This builds on top of #52199 by updating the dynamo backend to
implement backend.BackendWithItems. Both GetRange and DeleteRange
were refactored to call Items instead of getAllRecords to unify
logic and vet the implementation of Items.

The custom pagination logic to retrieve items was also replaced with
an iterator that works in a similar fashio to the query paginator
from the aws sdk. In addition to simplifying logic this also
removed some extraneous sorting.