[v15] Fix docs prose linting error for deleted files
Backports #52819

The docs prose linter currently fails when a pull request deletes a
file. This is because the `vale` CLI exits with a nonzero code when it
receives a nonexistent path. This change adds a filter to the
`dorny/paths-filter` job in addition to the `changed` filter that only
matches added or changed files. It passes this list of files to the
prose linter instead of the list of all changes files.