Re compute role.spec.options defaults every time instead of using the state value
The previous commits from Erik are fixing the role options default
computation. This is great but surfaces a new issue: our TF provider
doers not behave properly when default logic changes as it keeps using
the value from its state.

This commit makes Terraform re-compute the role.spec.options from the
original user config instead of the state. This makes sure that TF
options are the same regardless of which TF provider version created the
resource initially.